Made changes to existing mute and unmute
The primitive way of muting someone in discord, checks for certain roles, but that needs to be changed. Working on making sure the bot works on other servers in which users with roles can still use commands regardless. I've been testing as an owner so far, which testing on other servers where I'm not ran into some issues.